+++ ci/cpu/build.sh
@@ -5,13 +5,9 @@
 ################################################################################
 set -e
 
-# Logger function for build status output
-function logger() {
-  echo -e "\n>>>> $@\n"
-}
-
-# Set path
-export PATH=/conda/bin:/usr/local/cuda/bin:$PATH
+# Set path and build parallel level
+export PATH=/opt/conda/bin:/usr/local/cuda/bin:$PATH
+export PARALLEL_LEVEL=${PARALLEL_LEVEL:-4}
 
 # Set home to the job's workspace
 export HOME=$WORKSPACE
@@ -24,19 +20,30 @@ if [[ "$BUILD_MODE" = "branch" && "$SOURCE_BRANCH" = branch-* ]] ; then
   export VERSION_SUFFIX=`date +%y%m%d`
 fi
 
+# Setup 'gpuci_conda_retry' for build retries (results in 2 total attempts)
+export GPUCI_CONDA_RETRY_MAX=1
+export GPUCI_CONDA_RETRY_SLEEP=30
+
 ################################################################################
 # SETUP - Check environment
 ################################################################################
 
-logger "Get env..."
+gpuci_logger "Get env"
 env
 
-logger "Activate conda env..."
-source activate gdf
+gpuci_logger "Activate conda env"
+. /opt/conda/etc/profile.d/conda.sh
+conda activate rapids
 
-logger "Check versions..."
+gpuci_logger "Check versions"
 python --version
-conda list
+$CC --version
+$CXX --version
+
+gpuci_logger "Check conda environment"
+conda info
+conda config --show-sources
+conda list --show-channel-urls
 
 # FIX Added to deal with Anancoda SSL verification issues during conda builds
 conda config --set ssl_verify False
@@ -45,12 +52,12 @@ conda config --set ssl_verify False
 # BUILD - Conda package build
 ################################################################################
 
-logger "Build conda pkg for clx..."
-source ci/cpu/clx/build_clx.sh
+gpuci_logger "Build conda pkg for clx"
+conda build conda/recipes/clx
 
 ################################################################################
 # UPLOAD - Conda package
 ################################################################################
 
-logger "Upload clx conda pkg..."
-source ci/cpu/clx/upload-anaconda.sh
+gpuci_logger "Upload packages"
+source ci/cpu/upload.sh