+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## New Features
 
 ## Improvements
+- PR #276 Project Flash and gpuCI Scripts Update
 - PR #303 CLX sequence classifier
 - PR #308 CLX streamz workflow updates
 - PR #313 XGBoost training notebook